refactor(framework-tests) library classes renamed.
